requirements.txt
aiogram==2.16
aiohttp==3.8.0
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
et-xmlfile==1.1.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
mysql-connector-python==8.0.28
openpyxl==3.0.9
phonenumbers==8.12.41
protobuf==3.19.4
pytz==2021.3
PyYAML==6.0
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
xlwt==1.3.0
yarl==1.7.2